Header Id and Status are now CMDId and CMDStatus types.

+++ pdu_bind.go
@@ -104,8 +104,8 @@ func (s *Bind) writeFields() []byte {
 func (s *Bind) Writer() []byte {
 	b := s.writeFields()
 	h := packUi32(uint32(len(b) + 16))
-	h = append(h, packUi32(s.Header.Id)...)
-	h = append(h, packUi32(s.Header.Status)...)
+	h = append(h, packUi32(uint32(s.Header.Id))...)
+	h = append(h, packUi32(uint32(s.Header.Status))...)
 	h = append(h, packUi32(s.Header.Sequence)...)
 	return append(h, b...)
 }
