+++ pdu_enquire_link.go
@@ -63,8 +63,8 @@ func (s *EnquireLink) writeFields() []byte {
 func (s *EnquireLink) Writer() []byte {
 	b := s.writeFields()
 	h := packUi32(uint32(len(b) + 16))
-	h = append(h, packUi32(ENQUIRE_LINK)...)
-	h = append(h, packUi32(s.Header.Status)...)
+	h = append(h, packUi32(uint32(ENQUIRE_LINK))...)
+	h = append(h, packUi32(uint32(s.Header.Status))...)
 	h = append(h, packUi32(s.Header.Sequence)...)
 	return append(h, b...)
 }
